detekt_scanner/requirements.txt
TikTokApi>=7.3.0
google-cloud-pubsub>=2.25.0
google-cloud-firestore>=2.18.0
google-cloud-storage>=2.18.0
sightengine>=1.4.0
upstash-redis>=1.0.0
playwright>=1.40.0
httpx>=0.27.0
structlog>=24.0.0
